+++ src/Hoverfly.Core/Logging/ILoggerFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace Hoverfly.Core.Logging
-{
-    public interface ILoggerFactory
-    {
-        ILog Create(string name);
-    }
-}